Support disabled/readonly state

+++ src/components/Token.jsx
@@ -2,29 +2,25 @@ import React from 'react';
 
 export default class Token extends React.Component {
 
-	constructor (props) {
-		super(props);
-		this.remove = this.remove.bind(this);
-	}
-
 	static propTypes = {
 		onRemove: React.PropTypes.func,
 		value: React.PropTypes.string.isRequired
 	}
 
-	remove () {
+	remove = () => {
 		const {value, onRemove} = this.props;
-		onRemove(value);
+		if (onRemove) {
+			onRemove(value);
+		}
 	}
 
 	render () {
-
-		const {value} = this.props;
+		const {onRemove, value} = this.props;
 
 		return (
 			<div className="token">
 				<span className="value">{value}</span>
-				<i onClick={this.remove} className="icon-bold-x small"/>
+				{onRemove && ( <i onClick={this.remove} className="icon-bold-x small"/> )}
 			</div>
 		);
 	}
